Model now predicts te prospects
Got rid of a useless csv file, and made a copy of the nfl college stats csv. Changed the model to use the correct file path for the new csv file. Prospect_Predictions now uses the joblib model configuration to predict the scores of the prospects. CSV files were also changed to match eachother